backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-ratelimit>=3.0
dj-database-url>=2.1
psycopg2-binary>=2.9
